requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.21.0
Pillow>=8.3.1
onnxruntime>=1.8.0
python-multipart>=0.0.5